Fix: Resolved Currentuser class error by move to other space #8
