+++ .gitignore
@@ -56,3 +56,6 @@ temp/
 tmp/
 
 .copilot-instructions.md
+
+.idea/*
+.gradle/*
